Auslagern von Funktionen und Klassen
